+++ src/main/kotlin/cloud/glitchdev/rfu/data/fishing/HotspotCache.kt
@@ -83,4 +83,10 @@ object HotspotCache {
             }
         }
     }
+
+    fun clearSessionBuffs() {
+        synchronized(cache) {
+            cache.values.forEach { it.clearSession() }
+        }
+    }
 }