+++ CHANGELOG.md
@@ -2,8 +2,10 @@
 
 ### Features:
 - Removed 1.21.10 support
+- Added Hotspot Pointer with customizable priority.
+  - Use at your own risk! Off by default
 
-### Changes:
+### Changes: 
 
 ### Fixes: